requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.12.0
antlr4-tools==0.2
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
beautifulsoup4==4.12.0
blinker==1.4
bs4==0.0.1
cbor2==5.4.6
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
cryptography==3.4.8
dbus-python==1.2.18
decorator==4.4.2
distro==1.7.0
frozenlist==1.3.3
httplib2==0.20.2
idna==3.4
install-jdk==1.0.2
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==4.9.2
markdown-it-py==3.0.0
multidict==6.0.4
networkx==2.5.1
oauthlib==3.2.0
openai==0.27.8
packaging==23.0
prettytable==3.6.0
pycodestyle==2.10.0
pycryptodome==3.17
pygments=2.18.0
PyJWT==2.3.0
pysha3==1.0.2
PyYAML==6.0
requests==2.28.2
rich==13.7.1
six==1.16.0
soupsieve==2.4
tiktoken==0.4.0
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.6.3
urllib3==1.26.15
wadllib==1.3.6
wcwidth==0.2.6
yarl==1.8.2
z3-solver==4.11.2.0
zipp==1.0.0
falcon-analyzer@git+https://github.com/MetaTrustLabs/falcon-metatrust
